refactor(agent): Split tools into domain modules
tools.rs had grown past 2k lines with every tool, schema, and job helper
in one file. Split by capability (commands, patch, web, browser,
mandates, artifacts, skills, questions) behind a thin registry.
