+++ crates/sprocket-agent/README.md
@@ -53,7 +53,7 @@ flow back through the executor job. Released agents that still call the public
 - `provider.rs`: gateway completion loop, transcript sink, and provider outcomes.
 - `compaction.rs`: in-loop context compaction via a normal Responses call.
 - `history_json.rs`: Rig history JSON for the compaction summarizer.
-- `tools.rs`: model tools and durable job coordination.
+- `tools/`: model tools and durable job coordination.
 - `convex.rs`: run-control communication.
 - `types.rs`: history and context wire types.
 - `hooks.rs`: tool-call correlation, invalid-call handling, and OpenAI additional params.
