+++ crates/sprocket-agent/src/tools/web.rs
@@ -1,0 +1,116 @@
+use rig::tool::ToolExecutionError;
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+use serde_json::json;
+
+use super::context::{AgentToolContext, tool_error};
+use super::job::execute_cloud_tool_job;
+
+pub(super) const DEFAULT_WEB_SEARCH_RESULTS: u32 = 5;
+
+#[derive(Clone)]
+pub(crate) struct ScrapeUrlTool(pub(super) AgentToolContext);
+
+#[derive(Clone)]
+pub(crate) struct WebSearchTool(pub(super) AgentToolContext);
+
+fn default_web_search_results() -> u32 {
+    DEFAULT_WEB_SEARCH_RESULTS
+}
+
+fn is_default_web_search_results(num_results: &u32) -> bool {
+    *num_results == DEFAULT_WEB_SEARCH_RESULTS
+}
+
+pub(super) fn web_search_parameters() -> serde_json::Value {
+    let mut schema = json!(schemars::schema_for!(WebSearchArgs));
+    schema["properties"]["numResults"]["default"] = json!(DEFAULT_WEB_SEARCH_RESULTS);
+    schema
+}
+
+#[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
+pub(crate) struct WebSearchArgs {
+    /// Web search query.
+    pub(crate) query: String,
+    /// Number of results to return, between 1 and 10. Defaults to 5.
+    #[serde(
+        rename = "numResults",
+        default = "default_web_search_results",
+        skip_serializing_if = "is_default_web_search_results"
+    )]
+    #[schemars(default = "default_web_search_results")]
+    pub(crate) num_results: u32,
+}
+
+#[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
+pub(crate) struct ScrapeUrlArgs {
+    /// URL of the web page to read.
+    pub(crate) url: String,
+}
+
+
+impl rig::tool::Tool for WebSearchTool {
+    const NAME: &'static str = "web_search";
+    type Error = ToolExecutionError;
+    type Args = WebSearchArgs;
+    type Output = serde_json::Value;
+
+    fn description(&self) -> String {
+        "Search the web. Returns relevant pages with their URL, title, and a text excerpt. Use for current events or information beyond the local workspace."
+            .to_string()
+    }
+
+    fn parameters(&self) -> serde_json::Value {
+        web_search_parameters()
+    }
+
+    async fn call(
+        &self,
+        _context: &mut rig::tool::ToolContext,
+        args: Self::Args,
+    ) -> Result<Self::Output, Self::Error> {
+        let payload = serde_json::to_value(&args).map_err(|e| tool_error(e.into()))?;
+        execute_cloud_tool_job(
+            &self.0.runtime,
+            &self.0.run_id,
+            &self.0.claim_id,
+            Self::NAME,
+            &self.0.tool_call_tracker,
+            payload,
+        )
+        .await
+    }
+}
+
+impl rig::tool::Tool for ScrapeUrlTool {
+    const NAME: &'static str = "scrape_url";
+    type Error = ToolExecutionError;
+    type Args = ScrapeUrlArgs;
+    type Output = serde_json::Value;
+
+    fn description(&self) -> String {
+        "Read a web page by URL and return its content converted to markdown. Very long pages are truncated."
+            .to_string()
+    }
+
+    fn parameters(&self) -> serde_json::Value {
+        json!(schemars::schema_for!(ScrapeUrlArgs))
+    }
+
+    async fn call(
+        &self,
+        _context: &mut rig::tool::ToolContext,
+        args: Self::Args,
+    ) -> Result<Self::Output, Self::Error> {
+        let payload = serde_json::to_value(&args).map_err(|e| tool_error(e.into()))?;
+        execute_cloud_tool_job(
+            &self.0.runtime,
+            &self.0.run_id,
+            &self.0.claim_id,
+            Self::NAME,
+            &self.0.tool_call_tracker,
+            payload,
+        )
+        .await
+    }
+}
